feat: add top half of comment
