+++ src/index.css
@@ -17,7 +17,11 @@ li {
   list-style-type: none;
 }
 
-img {
+.padding10 {
+  padding: 10px;
+}
+
+#logo {
   width: 100px;
 }
 
@@ -100,17 +104,17 @@ img {
 }
 
 .modal {
-  display: none; /* Hidden by default */
-  position: fixed; /* Stay in place */
-  z-index: 1; /* Sit on top */
-  padding-top: 100px; /* Location of the box */
+  display: none;
+  position: fixed;
+  z-index: 1;
+  padding-top: 100px;
   left: 0;
   top: 0;
-  width: 100%; /* Full width */
-  height: 100%; /* Full height */
-  overflow: auto; /* Enable scroll if needed */
-  background-color: rgb(0, 0, 0); /* Fallback color */
-  background-color: rgba(0, 0, 0, 0.4); /* Black w/ opacity */
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgb(0, 0, 0);
+  background-color: rgba(0, 0, 0, 0.4);
 }
 
 .modal-content {
@@ -118,7 +122,7 @@ img {
   margin: auto;
   padding: 20px;
   border: 1px solid #888;
-  width: 80%;
+  width: 50%;
   display: flex;
   justify-content: space-between;
   align-content: flex-start;
@@ -136,4 +140,27 @@ img {
   color: #000;
   text-decoration: none;
   cursor: pointer;
+}
+
+.commentitem {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-content: center;
+  margin: 0 auto;
+}
+
+.commentitem img {
+  width: 80%;
+  align-self: center;
+}
+
+.commentitem-name {
+  align-self: center;
+}
+
+.commentitem-detail {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  align-self: center;
 }
