+++ src/index.html
@@ -25,7 +25,7 @@
     <div class="item-list">
       <div class="card-container"></div>
     </div>
-    <button id="myBtn">Open Modal</button>
+
     <div id="myModal" class="modal"></div>
 
     <footer class="footer">
