+++ src/modules/display-items.js
@@ -1,11 +1,12 @@
 import { cardContainer, logoContainer, modal } from './DOM-elements.js';
 import logo from '../assets/logo.jpg';
+import showComment from './comments.js';
 
 logoContainer.src = logo;
 
 function displayCards(data) {
   const imgUrl = data.sprites.other['official-artwork'].front_default;
-  const { name } = data;
+  const { name, weight, height } = data;
   const card = document.createElement('div');
   card.dataset.name = name;
   card.classList.add('card');
@@ -17,10 +18,21 @@ function displayCards(data) {
           <div class="likes">5 likes</div>
           <button type='button' class="comments">Comments</button>
           <button type='button' class="reservations">Reservations</button>`;
+  card.children[3].addEventListener('click', () => {
+    showComment(name, imgUrl, height, weight, data.types[0].type.name, data.moves[0].move.name);
+    const span = document.createElement('span');
+    span.innerHTML = '&times;';
+    span.classList.add('close');
+    span.addEventListener('click', () => {
+      modal.style.display = 'none';
+    });
+    modal.children[0].append(span);
+    modal.style.display = 'block';
+  });
   cardContainer.appendChild(card);
 }
 
-export function displayChar(name) {
+export default function displayChar(name) {
   const url = `https://pokeapi.co/api/v2/pokemon/${name}`;
   fetch(url)
     .then((response) => response.json())
@@ -29,12 +41,3 @@ export function displayChar(name) {
     })
     .catch((err) => `Pokemon not found, ${err}`);
 }
-
-export const showComment = () => {
-  const comment = `
-  <div class="modal-content">
-    <p>Some text in the Modal..</p>
-  </div>
-`;
-  modal.innerHTML = comment;
-};
